+++ spb-code-base/src/main/java/prm/projectbase/dto/request/UserUpdateRequest.java
@@ -28,5 +28,5 @@ public class UserUpdateRequest {
 
     Boolean active;
 
-    Integer roleId;
+    Long roleId;
 }
