+++ spb-code-base/src/main/java/prm/projectbase/repository/ClassroomRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
-    List<Classroom> findByTeacherIdAndActiveTrue(Integer teacherId);
+    List<Classroom> findByTeacherIdAndActiveTrue(Long teacherId);
     List<Classroom> findByActiveTrue();
     Optional<Classroom> findByJoinCode(String joinCode);
     boolean existsByCode(String code);
