+++ spb-code-base/src/main/java/prm/projectbase/repository/PeerReviewRepository.java
@@ -6,7 +6,7 @@ import prm.projectbase.entity.PeerReview;
 import java.util.List;
 import java.util.Optional;
 
-public interface PeerReviewRepository extends JpaRepository<PeerReview, Integer> {
-    List<PeerReview> findByClassroomIdAndReviewerStudentId(Integer classroomId, Integer reviewerStudentId);
-    Optional<PeerReview> findByReviewerStudentIdAndReviewedGroupId(Integer reviewerStudentId, Integer reviewedGroupId);
+public interface PeerReviewRepository extends JpaRepository<PeerReview, Long> {
+    List<PeerReview> findByClassroomIdAndReviewerStudentId(Long classroomId, Long reviewerStudentId);
+    Optional<PeerReview> findByReviewerStudentIdAndReviewedGroupId(Long reviewerStudentId, Long reviewedGroupId);
 }
